Remove duplicate changeRequest functionality in API and move the tests to the changeRequest rest endpoint
